Restrict snapshot file permissions
